+++ test/new-e2e/pkg/environments/aws/docker/host.go
@@ -51,6 +51,7 @@ func newProvisionerParams() *ProvisionerParams {
 	}
 }
 
+// GetProvisionerParams return ProvisionerParams from options opts setup
 func GetProvisionerParams(opts ...ProvisionerOption) *ProvisionerParams {
 	params := newProvisionerParams()
 	err := optional.ApplyOptions(params, opts)
