+++ test/new-e2e/tests/npm/ec2_1host_test.go
@@ -6,15 +6,10 @@
 package npm
 
 import (
-	"math"
-	"os"
 	"testing"
-	"time"
 
-	agentmodel "github.com/DataDog/agent-payload/v5/process"
 	"github.com/pulumi/pulumi/sdk/v3/go/pulumi"
 
-	"github.com/DataDog/datadog-agent/test/fakeintake/aggregator"
 	"github.com/DataDog/datadog-agent/test/new-e2e/pkg/components"
 	"github.com/DataDog/datadog-agent/test/new-e2e/pkg/e2e"
 	"github.com/DataDog/datadog-agent/test/new-e2e/pkg/environments"
@@ -24,8 +19,6 @@ import (
 	"github.com/DataDog/test-infra-definitions/components/docker"
 	"github.com/DataDog/test-infra-definitions/resources/aws"
 	"github.com/DataDog/test-infra-definitions/scenarios/aws/ec2"
-
-	"github.com/stretchr/testify/assert"
 )
 
 type hostHttpbinEnv struct {
@@ -49,7 +42,7 @@ func hostDockerHttpbinEnvProvisioner() e2e.PulumiEnvRunFunc[hostHttpbinEnv] {
 			awshost.WithAgentOptions(agentparams.WithSystemProbeConfig(systemProbeConfigNPM)),
 		}
 		params := awshost.GetProvisionerParams(opts...)
-		awshost.HostRunFunction(ctx, &env.Host, params)
+		awshost.Run(ctx, &env.Host, params)
 
 		vmName := "httpbinvm"
 
@@ -83,10 +76,6 @@ func TestEC2VMSuite(t *testing.T) {
 	s := &ec2VMSuite{}
 
 	e2eParams := []e2e.SuiteOption{e2e.WithProvisioner(e2e.NewTypedPulumiProvisioner("hostHttpbin", hostDockerHttpbinEnvProvisioner(), nil))}
-	// debug helper
-	if _, devmode := os.LookupEnv("TESTS_E2E_DEVMODE"); devmode {
-		e2eParams = append(e2eParams, e2e.WithDevMode())
-	}
 
 	// Source of our kitchen CI images test/kitchen/platforms.json
 	// Other VM image can be used, our kitchen CI images test/kitchen/platforms.json
@@ -98,11 +87,16 @@ func (v *ec2VMSuite) SetupSuite() {
 	v.BaseSuite.SetupSuite()
 
 	v.Env().RemoteHost.MustExecute("sudo apt install -y apache2-utils")
+	v.Env().RemoteHost.MustExecute("sudo snap install docker")
 }
 
 // BeforeTest will be called before each test
 func (v *ec2VMSuite) BeforeTest(suiteName, testName string) {
 	v.BaseSuite.BeforeTest(suiteName, testName)
+	v.beforeTest(suiteName, testName)
+}
+
+func (v *ec2VMSuite) beforeTest(suiteName, testName string) {
 	// default is to reset the current state of the fakeintake aggregators
 	if !v.BaseSuite.IsDevMode() {
 		v.Env().FakeIntake.Client().FlushServerAndResetAggregators()
@@ -110,47 +104,30 @@ func (v *ec2VMSuite) BeforeTest(suiteName, testName string) {
 }
 
 // TestFakeIntakeNPM Validate the agent can communicate with the (fake) backend and send connections every 30 seconds
+// 2 tests generate the request on the host and on docker
 //   - looking for 1 host to send CollectorConnections payload to the fakeintake
 //   - looking for 3 payloads and check if the last 2 have a span of 30s +/- 500ms
 func (v *ec2VMSuite) TestFakeIntakeNPM() {
-	t := v.T()
+	vt := v.T()
 	testURL := "http://" + v.Env().HTTPBinHost.Address + "/"
+	vt.Run("host_requests", func(t *testing.T) {
+		v.BaseSuite.SetT(t)
+		v.beforeTest("TestEC2VMSuite", t.Name()) // workaround as suite doesn't call BeforeTest before each sub tests
 
-	// generate a connection
-	v.Env().RemoteHost.MustExecute("curl " + testURL)
+		// generate a connection
+		v.Env().RemoteHost.MustExecute("curl " + testURL)
 
-	targetHostnameNetID := ""
-	// looking for 1 host to send CollectorConnections payload to the fakeintake
-	v.EventuallyWithT(func(c *assert.CollectT) {
+		test1HostFakeIntakeNPM(&v.BaseSuite, v.Env().FakeIntake)
+	})
+	vt.Run("docker_requests", func(t *testing.T) {
+		v.BaseSuite.SetT(t)
+		v.beforeTest("TestEC2VMSuite", t.Name()) // workaround as suite doesn't call BeforeTest before each sub tests
 
-		hostnameNetID, err := v.Env().FakeIntake.Client().GetConnectionsNames()
-		assert.NoError(c, err, "GetConnectionsNames() errors")
-		if !assert.NotEmpty(c, hostnameNetID, "no connections yet") {
-			return
-		}
-		targetHostnameNetID = hostnameNetID[0]
+		// generate a connection
+		v.Env().RemoteHost.MustExecute("docker run curlimages/curl curl " + testURL)
 
-		t.Logf("hostname+networkID %v seen connections", hostnameNetID)
-	}, 60*time.Second, time.Second, "no connections received")
-
-	// looking for 3 payloads and check if the last 2 have a span of 30s +/- 500ms
-	v.EventuallyWithT(func(c *assert.CollectT) {
-		cnx, err := v.Env().FakeIntake.Client().GetConnections()
-		assert.NoError(t, err)
-
-		if !assert.Greater(c, len(cnx.GetPayloadsByName(targetHostnameNetID)), 2, "not enough payloads") {
-			return
-		}
-		var payloadsTimestamps []time.Time
-		for _, cc := range cnx.GetPayloadsByName(targetHostnameNetID) {
-			payloadsTimestamps = append(payloadsTimestamps, cc.GetCollectedTime())
-		}
-		dt := payloadsTimestamps[2].Sub(payloadsTimestamps[1]).Seconds()
-		t.Logf("hostname+networkID %v diff time %f seconds", targetHostnameNetID, dt)
-
-		// we want the test fail now, not retrying on the next payloads
-		assert.Greater(t, 0.5, math.Abs(dt-30), "delta between collection is higher than 500ms")
-	}, 90*time.Second, time.Second, "not enough connections received")
+		test1HostFakeIntakeNPM(&v.BaseSuite, v.Env().FakeIntake)
+	})
 }
 
 // TestFakeIntakeNPM_600cnx_bucket Validate the agent can communicate with the (fake) backend and send connections
@@ -158,115 +135,51 @@ func (v *ec2VMSuite) TestFakeIntakeNPM() {
 //   - looking for 1 host to send CollectorConnections payload to the fakeintake
 //   - looking for n payloads and check if the last 2 have a maximum span of 100ms
 func (v *ec2VMSuite) TestFakeIntakeNPM_600cnx_bucket() {
-	t := v.T()
+	vt := v.T()
 	testURL := "http://" + v.Env().HTTPBinHost.Address + "/"
+	vt.Run("host_requests", func(t *testing.T) {
+		v.BaseSuite.SetT(t)
+		v.beforeTest("TestEC2VMSuite", t.Name()) // workaround as suite doesn't call BeforeTest before each sub tests
 
-	// generate connections
-	v.Env().RemoteHost.MustExecute("ab -n 600 -c 600 " + testURL)
+		// generate connections
+		v.Env().RemoteHost.MustExecute("ab -n 600 -c 600 " + testURL)
 
-	targetHostnameNetID := ""
-	// looking for 1 host to send CollectorConnections payload to the fakeintake
-	v.EventuallyWithT(func(c *assert.CollectT) {
-		hostnameNetID, err := v.Env().FakeIntake.Client().GetConnectionsNames()
-		assert.NoError(c, err, "GetConnectionsNames() errors")
-		if !assert.NotEmpty(c, hostnameNetID, "no connections yet") {
-			return
-		}
-		targetHostnameNetID = hostnameNetID[0]
+		test1HostFakeIntakeNPM600cnxBucket(&v.BaseSuite, v.Env().FakeIntake)
+	})
+	vt.Run("docker_requests", func(t *testing.T) {
+		v.BaseSuite.SetT(t)
+		v.beforeTest("TestEC2VMSuite", t.Name()) // workaround as suite doesn't call BeforeTest before each sub tests
 
-		t.Logf("hostname+networkID %v seen connections", hostnameNetID)
-	}, 60*time.Second, time.Second, "no connections received")
+		// generate connections
+		v.Env().RemoteHost.MustExecute("docker run devth/alpine-bench -n 600 -c 600 " + testURL)
 
-	// looking for x payloads (with max 600 connections) and check if the last 2 have a max span of 100ms
-	v.EventuallyWithT(func(c *assert.CollectT) {
-		cnx, err := v.Env().FakeIntake.Client().GetConnections()
-		assert.NoError(t, err)
-
-		if !assert.GreaterOrEqualf(c, len(cnx.GetPayloadsByName(targetHostnameNetID)), 2, "not enough payloads") {
-			return
-		}
-
-		cnx.ForeachHostnameConnections(func(cnx *aggregator.Connections, hostname string) {
-			assert.LessOrEqualf(t, len(cnx.Connections), 600, "too many payloads")
-		})
-
-		hostPayloads := cnx.GetPayloadsByName(targetHostnameNetID)
-		lenHostPayloads := len(hostPayloads)
-		if !assert.Equalf(c, len(hostPayloads[lenHostPayloads-2].Connections), 600, "can't found enough connections 600+") {
-			return
-		}
-
-		cnx600PayloadTime := hostPayloads[lenHostPayloads-2].GetCollectedTime()
-		latestPayloadTime := hostPayloads[lenHostPayloads-1].GetCollectedTime()
-
-		dt := latestPayloadTime.Sub(cnx600PayloadTime).Seconds()
-		t.Logf("hostname+networkID %v diff time %f seconds", targetHostnameNetID, dt)
-
-		assert.Greater(t, 0.1, dt, "delta between collection is higher than 100ms")
-	}, 90*time.Second, time.Second, "not enough connections received")
+		test1HostFakeIntakeNPM600cnxBucket(&v.BaseSuite, v.Env().FakeIntake)
+	})
 }
 
 // TestFakeIntakeNPM_TCP_UDP_DNS validate we received tcp, udp, and DNS connections
 // with some basic checks, like IPs/Ports present, DNS query has been captured, ...
 func (v *ec2VMSuite) TestFakeIntakeNPM_TCP_UDP_DNS() {
-	t := v.T()
+	vt := v.T()
 	testURL := "http://" + v.Env().HTTPBinHost.Address + "/"
+	vt.Run("host_requests", func(t *testing.T) {
+		v.BaseSuite.SetT(t)
+		v.beforeTest("TestEC2VMSuite", t.Name()) // workaround as suite doesn't call BeforeTest before each sub tests
 
-	// generate connections
-	v.Env().RemoteHost.MustExecute("curl " + testURL)
-	v.Env().RemoteHost.MustExecute("dig @8.8.8.8 www.google.ch")
+		// generate connections
+		v.Env().RemoteHost.MustExecute("curl " + testURL)
+		v.Env().RemoteHost.MustExecute("dig @8.8.8.8 www.google.ch")
 
-	v.EventuallyWithT(func(c *assert.CollectT) {
+		test1HostFakeIntakeNPMTCPUDPDNS(&v.BaseSuite, v.Env().FakeIntake)
+	})
+	vt.Run("docker_requests", func(t *testing.T) {
+		v.BaseSuite.SetT(t)
+		v.beforeTest("TestEC2VMSuite", t.Name()) // workaround as suite doesn't call BeforeTest before each sub tests
 
-		cnx, err := v.Env().FakeIntake.Client().GetConnections()
-		assert.NoError(c, err, "GetConnections() errors")
-		if !assert.NotEmpty(c, cnx.GetNames(), "no connections yet") {
-			return
-		}
+		// generate connections
+		v.Env().RemoteHost.MustExecute("docker run curlimages/curl curl " + testURL)
+		v.Env().RemoteHost.MustExecute("docker run makocchi/alpine-dig dig @8.8.8.8 www.google.ch")
 
-		foundDNS := false
-		cnx.ForeachConnection(func(c *agentmodel.Connection, cc *agentmodel.CollectorConnections, hostname string) {
-			if len(c.DnsStatsByDomainOffsetByQueryType) > 0 {
-				foundDNS = true
-			}
-		})
-		if !assert.True(c, foundDNS, "DNS not found") {
-			return
-		}
-
-		type countCnx struct {
-			hit int
-			TCP int
-			UDP int
-		}
-		countConnections := make(map[string]*countCnx)
-
-		helperCleanup(t)
-		cnx.ForeachConnection(func(c *agentmodel.Connection, cc *agentmodel.CollectorConnections, hostname string) {
-			var count *countCnx
-			var found bool
-			if count, found = countConnections[hostname]; !found {
-				countConnections[hostname] = &countCnx{}
-				count = countConnections[hostname]
-			}
-			count.hit++
-
-			switch c.Type {
-			case agentmodel.ConnectionType_tcp:
-				count.TCP++
-			case agentmodel.ConnectionType_udp:
-				count.UDP++
-			}
-			validateConnection(t, c, cc, hostname)
-		})
-
-		totalConnections := countCnx{}
-		for host, count := range countConnections {
-			t.Logf("connections %d tcp %d udp %d received by host/netID %s", count.hit, count.TCP, count.UDP, host)
-			totalConnections.hit += count.hit
-			totalConnections.TCP += count.TCP
-			totalConnections.UDP += count.UDP
-		}
-		t.Logf("sum connections %d tcp %d udp %d", totalConnections.hit, totalConnections.TCP, totalConnections.UDP)
-	}, 60*time.Second, time.Second)
+		test1HostFakeIntakeNPMTCPUDPDNS(&v.BaseSuite, v.Env().FakeIntake)
+	})
 }
